Desktop/Trabalho Rafael/criptobot/requirements.txt
praw
vaderSentiment
requests
python-dotenv
